Inclusion of omit_resources context flag for restricted_package_show
It is sometimes not ideal for the result of a package_show call to parse all of its resources (i.e. if we're only interested in the general metadata that the package represents). A context flag called `omit_resources` is paid attention to in `restricted_package_show`, which if true, will not parse the resources and simply remove them from the returned dictionary.