For particular endpoints, a "lite" representation of resources will be returned for restricted_package_show
There are some Blueprint views in which a dataset's resource information doesn't need to be evaluated in full for restriction checking. These include pages which use the /snippets/package_list.html snippet. A list of endpoints has been defined in logic.py which the current request.endpoint is checked against. If the request is for one of those endpoints, the package resources are reduced to a subset of "safe" properties, which don't require any redacting/hiding if the current user doesn't necessarily have the right permissions to view the resource in full.